🛠️ :: RecruitAreaResponse를 RecruitDetailsResponse의 이너클래스로 변경
