+++ src/main/java/com/example/jobis/domain/recruit/controller/dto/response/RecruitDetailsResponse.java
@@ -1,5 +1,8 @@
 package com.example.jobis.domain.recruit.controller.dto.response;
 
+import com.example.jobis.domain.code.domain.Code;
+import com.example.jobis.domain.code.domain.RecruitAreaCode;
+import com.example.jobis.domain.code.domain.enums.CodeType;
 import com.example.jobis.domain.recruit.domain.RecruitArea;
 import com.example.jobis.domain.recruit.domain.enums.ProgressType;
 import lombok.Builder;
@@ -7,6 +10,7 @@ import lombok.Getter;
 
 import java.time.LocalDate;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Getter
 @Builder
@@ -28,4 +32,36 @@ public class RecruitDetailsResponse {
     private LocalDate endDate;
     private String etc;
 
+    @Getter
+    @Builder
+    public static class RecruitAreaResponse {
+        private List<String> job;
+        private List<String> tech;
+        private int hiring;
+        private String majorTask;
+
+        public static RecruitAreaResponse of(RecruitArea recruitArea) {
+
+            List<String> jobCodes = recruitArea.getCodeList().stream()
+                    .map(RecruitAreaCode::getCodeId)
+                    .toList().stream()
+                    .filter(code -> code.getCodeType().equals(CodeType.JOB))
+                    .map(Code::getKeyword)
+                    .collect(Collectors.toList());
+
+            List<String> techCodes = recruitArea.getCodeList().stream()
+                    .map(RecruitAreaCode::getCodeId)
+                    .toList().stream()
+                    .filter(code -> code.getCodeType().equals(CodeType.TECH))
+                    .map(Code::getKeyword)
+                    .collect(Collectors.toList());
+
+            return RecruitAreaResponse.builder()
+                    .job(jobCodes)
+                    .tech(techCodes)
+                    .hiring(recruitArea.getHiredCount())
+                    .majorTask(recruitArea.getMajorTask())
+                    .build();
+        }
+    }
 }
